consolidate attack and bitboards
