+++ src/evaluate.c
@@ -18,7 +18,6 @@
 
 #include <stdlib.h>
 
-#include "attack.h"
 #include "bitboards.h"
 #include "board.h"
 #include "evaluate.h"
