+++ src/makemove.c
@@ -16,7 +16,6 @@
   along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#include "attack.h"
 #include "bitboards.h"
 #include "board.h"
 #include "psqt.h"
